Expression Worksheet: Dog Years
Created a variable for Sparky’s age, a variable for human to dog years,
and a variable for the result. In the dogYrs variable I created an
expression in which I concatenated the required strings and multiplied
Sparky’s human years by the human-to-dog years. I then used the
console.log function to print the concatenated string in the dogYrs
variable to the console.