Expression Worksheet: Edits and Slice of Pie
Added some more comments to the first and second problem. Calculated
the second problem (number of slices times number of pizzas divided by
the number of people) and concatenated the strings. Stored the result
in the numSlices variable, and used the console.log function to write
it to the console.