Expression Worksheet: Average Shopping Bill
Declared and defined the variable shopTotals with an array to hold all
five shopping totals, then declared and defined the totalAmt variable
to add all five elements together, then declared and defined the
average variable and calculated the average of the five week’s bills.
Concatenated the strings in the console.log function and wrote the
result to the console.